now clicking the label does not delete the item

+++ client/src/App.js
@@ -1,23 +1,26 @@
 import React, { useState, useEffect, useCallback } from "react";
 import TextField from "@mui/material/TextField";
+import Typography from "@mui/material/Typography";
 import FormGroup from "@mui/material/FormGroup";
-import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
 import "./App.css";
 
 function NewItem({ setNeedsUpdate }) {
   const [item, setItem] = useState("");
 
-  const saveNewItem = useCallback(async (item) => {
-    const requestOptions = {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ item }),
-    };
-    await fetch("/api/add_recurring_item", requestOptions);
-    setNeedsUpdate(true);
-    setItem("")
-  }, [setNeedsUpdate, setItem]);
+  const saveNewItem = useCallback(
+    async (item) => {
+      const requestOptions = {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ item }),
+      };
+      await fetch("/api/add_recurring_item", requestOptions);
+      setNeedsUpdate(true);
+      setItem("");
+    },
+    [setNeedsUpdate, setItem]
+  );
 
   return (
     <span className="NewItem">
@@ -30,7 +33,6 @@ function NewItem({ setNeedsUpdate }) {
           onKeyPress: async (event) => {
             if (event.key === "Enter") {
               event.preventDefault();
-              console.log("sundra1", item);
               await saveNewItem(item);
             }
           },
@@ -40,29 +42,39 @@ function NewItem({ setNeedsUpdate }) {
   );
 }
 
-function ExistingItemList({ items, setNeedsUpdate }) {
-  const handleCheck = useCallback(async (event, item) => {
-    if (event.target.checked) {
-      console.log("sundra", item);
-      const requestOptions = {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ item }),
-      };
-      await fetch("/api/remove_recurring_item", requestOptions);
-      setNeedsUpdate(true);
-    }
-  }, [setNeedsUpdate]);
+function ExistingItem({ item, setNeedsUpdate }) {
+  const handleCheck = useCallback(
+    async (event) => {
+      if (event.target.checked && event.target === event.currentTarget) {
+        console.log(event);
+        const requestOptions = {
+          method: "DELETE",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ item }),
+        };
+        await fetch("/api/remove_recurring_item", requestOptions);
+        setNeedsUpdate(true);
+      }
+    },
+    [setNeedsUpdate, item]
+  );
 
+  return (
+    <span className="ExistingItem">
+      <Typography gutterBottom={false}>
+        {" "}
+        <Checkbox onChange={async (e) => await handleCheck(e)} />
+        {item}
+      </Typography>
+    </span>
+  );
+}
+
+function ExistingItemList({ items, setNeedsUpdate }) {
   return (
     <FormGroup>
       {items.map((item, i) => (
-        <FormControlLabel
-          // TODO (sandrasandeep): prevent onChange when label, not checkbox, is clicked
-          control={<Checkbox onChange={async (e) => await handleCheck(e, item)} />}
-          label={item}
-          key={i}
-        />
+        <ExistingItem key={i} item={item} setNeedsUpdate={setNeedsUpdate} />
       ))}
     </FormGroup>
   );
@@ -75,23 +87,25 @@ function App() {
   useEffect(() => {
     if (needsUpdate) {
       fetch("/api/get_recurring_list")
-      .then((result) => result.json())
-      .then((result) => {
-        setRecurringList(result);
-      });
+        .then((result) => result.json())
+        .then((result) => {
+          setRecurringList(result);
+        });
       setNeedsUpdate(false);
     }
   }, [needsUpdate, setNeedsUpdate]);
 
   return (
-    // TODO(sandra-sandeep): add weekly (non-recurring) grocery list too.
     <div>
       {typeof recurringList === "undefined" ? (
         <p>Loading...</p>
       ) : (
         // TODO(sandra-sandeep): Make it cute.
         <div>
-          <ExistingItemList items={recurringList} setNeedsUpdate={setNeedsUpdate} />
+          <ExistingItemList
+            items={recurringList}
+            setNeedsUpdate={setNeedsUpdate}
+          />
           <NewItem setNeedsUpdate={setNeedsUpdate} />
         </div>
       )}
